# docs/src/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Message Vault HTTP API",
    "description": "HTTP API for a local Message Vault. Bearer session tokens come from login. API tokens come from Settings → Account.",
    "license": {
      "name": "Fair Core License 1.0 (ALv2 future)",
      "url": "https://github.com/bitrealm-io/message-vault/blob/main/LICENSE.md"
    },
    "version": "0.8.3"
  },
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a local vault account and return its session token.",
        "operationId": "register_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Public registration is off",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/owner/vault-settings": {
      "get": {
        "tags": [
          "Owner"
        ],
        "summary": "Read the vault's settings.",
        "operationId": "owner_vault_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Owner"
        ],
        "summary": "Change the vault's settings.",
        "operationId": "owner_patch_vault_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchVaultSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/vault/claim": {
      "post": {
        "tags": [
          "Vault"
        ],
        "summary": "Claim an unclaimed vault by creating its owner.",
        "description": "Unauthenticated, because a vault with no owner has no credential that\ncould authorize this. Whoever reaches an unclaimed vault first may claim\nit: the vault is self-hosted, so its operator installs the software,\nclaims the vault, and publishes the port, in that order and at times of\ntheir choosing. An unclaimed vault is also empty, so a lost race destroys\nnothing and announces itself at once.",
        "operationId": "claim_vault",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimVaultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vault claimed; session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Already claimed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthTokenResponse": {
        "type": "object",
        "description": "Session token plus the account id and username it belongs to.",
        "required": [
          "token",
          "account_id",
          "username"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "description": "Account id the session belongs to."
          },
          "token": {
            "type": "string",
            "description": "Session token to send as `Authorization: Bearer …`."
          },
          "username": {
            "type": "string",
            "description": "Account username (falls back to the account id)."
          }
        }
      },
...
      "ClaimVaultRequest": {
        "type": "object",
        "description": "Body for claiming a vault.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "Password for the vault owner. Must satisfy the vault's password policy."
          },
          "username": {
            "type": "string",
            "description": "Login username for the vault owner."
          }
        }
      },
...
      "PatchVaultSettingsRequest": {
        "type": "object",
        "description": "Body for changing the vault's settings. Omitted fields are left alone.",
        "properties": {
          "public_registration": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Let anyone reaching the vault create their own account, or stop them."
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "An RFC 7807 problem document: the body of every failure the vault answers,\nserved as `application/problem+json` (ADR-0010).\n\n`type` is the URL of the page describing this kind of failure, one page per\ntype, or `about:blank` for an internal error. A validation failure carries\n`errors`, every field that failed, in place of `detail`. Every problem\nrepeats the response's `x-request-id` as `request_id`, so the person\nreading the failure and the operator reading the log are looking at the\nsame request.\n\nThe extension members belong to one type each: `word` and `did_you_mean`\nto `search-query-invalid`, `retry_after` to `rate-limited`.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "One sentence about this occurrence. Absent on a validation failure,\nwhich lists `errors` instead."
          },
          "did_you_mean": {
            "type": [
              "string",
              "null"
            ],
            "description": "`search-query-invalid`: a word the language does have, within a small\nedit distance of the one used."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every rule the request broke, one sentence each. Only on\n`validation-failed`."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `x-request-id` of the response this came in."
          },
          "retry_after": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "`rate-limited`: seconds until an attempt may succeed, the same figure\nas the `Retry-After` header.",
            "minimum": 0
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status, repeated in the body.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The type's fixed, human-readable name."
          },
          "type": {
            "type": "string",
            "description": "URL of the page describing this kind of failure; `about:blank` for a\n`500 Internal Server Error`."
          },
          "word": {
            "type": [
              "string",
              "null"
            ],
            "description": "`search-query-invalid`: the `word:` the query used."
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Body for local account registration.",
        "required": [
          "username"
        ],
        "properties": {
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local password; absent or empty registers an account without one."
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "Phone number linked to the account."
          },
          "preferred_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name shown in the vault."
          },
          "username": {
            "type": "string",
            "description": "Login username."
          }
        }
      },
...
      "VaultSettingsResponse": {
        "type": "object",
        "description": "The vault settings the owner controls.",
        "required": [
          "public_registration"
        ],
        "properties": {
          "public_registration": {
            "type": "boolean",
            "description": "Anyone reaching the vault may create their own account."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}